added bimodal and minimum congestion status
Bimodal congestion status is a modification to BufferOccupancy and
PhantomBufferOccupancy as a new mode. It performs the following
function to blend the VC and port oriented information:
Q = Qvc * Qvc + (1 - Qvc) * Qport

Minimum congestion status is a simply value programmed into the top
level congestion status module. It simply doesn't allow the Q value
to drop below a certain value. This is applied after the granularization.
This is useful for weighted routing algorithms where a 0 value caused
non-minimal paths to look overly good.